Update switch.py
PORT GROUPS:

A port group serves as a pseudo layer2 network. Where ports in the group can send messages to other ports in the same group.

If a port is not in a group it will rely on an access-list. If a port is in a group it will not be able to communicate with: ports not in the same group or ports not in a group. 

A port can only be in one network at a time. 

Once again this code is very rough. Please bear with me!!