Update switch.py
++ADDED++

More functions for interacting with ports. It should also be noted port based functions require a user to authenticate before the request is accepted. 

-> client can change key for owned port
-> client can set description for owned port


The modularity of this code allows me to easily add new functions. Since we 1st, grab the data we need for a request and then pass it to functions that will modify the data.

Starting off, when a port needs a change, for each change, the database would need to be queried. This creates unnecessary performance overhead. Currently, when a port needs a change, all changes are made before committing  into the database.

Programming is virtually Lego building blocks!