***UPDATING***  function -> Send_Data
This function has been updated to successfully send data to another registered user. If  source and destination ports are not in groups then we must fallback to using the access list. The access list for a port is managed by the owner. Who may add or remove a user from this access list.

Therefore, we only have two modes for sending data.

Mode 1, if two ports are in the same group they may transmit data bidirectionally.

Mode 2, if two ports are not in a group, but have entries in the access list allowing permitting the client(s). 

The switch drops all other conditions since they are not valid.