Implementing Acks and Ack based functions
This part was a bit hard to wrap my head around due to being a silly human. The main issue was ensuring that the logic works correctly if an ack is needed or not. 

+ Acknowledgements are optional and only enforced by the source(sender) of the message. Which requests that a recipient (destination) acknowledges they received the data.  

Once an ack message is created it becomes locked. The owner of the ack may not delete the ack or alter any states. Instead, the receiver must acknowledge first. 

Then the sender can ack that the destination ack'd the message. Therefore confirming the receiver received the message. 

Acks works like a gated passage way. Here's an example:

Bob sends a message which has the ack flag set to  "TRUE" for Jen 

An ack message is made on the switch, for that message, with the owner being Bob and the responder is Jen.

Bob cannot alter this ack or delete it since we are awaiting Jens response.

Jen responds to the ack
Bob can now ack that Jen ack'd
The ack is deleted because Bob was able to confirm Jen received the message. 

ACK ACK ACK sounds like someone laughing


There may be future improvements for this function