Logic Additions
Added some more logic for keep track of stats for a port. Previously there were a couple of issues:

Even if a port didn't have access to send data to another port, it would still count as a transmit and the destination would count the receive. 

We don't want this type of behavior because the message shouldn't be accounted for when it's rejected.