This script will handle all things rate related
It's important for an API to have rate limits in place to keep from overloading the system. Rate limits will be applied on a per user basis. Regardless of the call made, the frequency between the next call is monitored. 

A user will be limited to 30 calls per minute or one call every 2 seconds.